### 2023-04-12 #97 @billbrod
Presubmission inquiry for plenoptic
Submitting Author: @billbrod

Package Name: plenoptic

One-Line Description of Package: a python library for model-based image synthesis.

Repository Link (if existing):   https://github.com/LabForComputationalVision/plenoptic/ ([this PR docs build](https://plenoptic--201.org.readthedocs.build/en/201/) may be more useful for an introduction, still haven't merged into main yet)

---

- Include a brief paragraph describing what your package does: `plenoptic` provides tools to help researchers understand their model by synthesizing novel informative stimuli, which help build intuition for what features the model ignores and what it is sensitive to.

We partner with communities to support peer review with an additional layer of
checks that satisfy community requirements. If your package fits into an

I think `plenoptic` is actually out of scope, but I wanted to check, because pyopensci looks cool. This package is intended for use by the vision science, machine learning, and neuroscience communities, but could be used by any researcher that builds models that take something image-, video-, or audio-like as input. The package generates new stimuli (for use in further experiments) rather than facilitates the visualization of existing data.

Researchers in vision science, machine learning, and neuroscience, largely.  The goal is to generate novel stimuli (images, videos, audio) that researchers can use in new experiments to better understand their computational models.

Not that I'm aware of.

[Conduct]: https://www.pyopensci.org/governance/CODE_OF_CONDUCT
