### 2023-04-12 90607c9
revises simulate/fit/synth description
