### 2023-04-13 84c3956
small change to allowed_methods
