### 2023-04-13 a66de22
fixes failing docstrings
these were largely mispellings of actual titles. sphinx used to silently
fail to render them correctly, but was now throwing an error
