+++ plenoptic/simulate/canonical_computations/non_linearities.py
@@ -102,8 +102,8 @@ def local_gain_control(x, epsilon=1e-8):
         The local phase of ``x`` (aka. local unit vector, or local
         state)
 
-    Note
-    ----
+    Notes
+    -----
     This function is an analogue to rectangular_to_polar for
     real valued signals.
     Norm and direction (analogous to complex modulus and phase) are
@@ -145,8 +145,8 @@ def local_gain_release(norm, direction, epsilon=1e-8):
     x : torch.Tensor
         Tensor of shape (B,C,H,W)
 
-    Note
-    ----
+    Notes
+    -----
     This function is an analogue to polar_to_rectangular for
     real valued signals.
     Norm and direction (analogous to complex modulus and phase) are
@@ -185,8 +185,9 @@ def local_gain_control_dict(coeff_dict, residuals=True):
     state: dict
         The dictionary of torch.Tensors containing the local phase of
         ``x``.
-    Note
-    ----
+
+    Notes
+    -----
     Note that energy and state is not computed on the residuals.
 
     The inverse operation is achieved by `local_gain_release_dict`.
@@ -228,8 +229,8 @@ def local_gain_release_dict(energy, state, residuals=True):
     coeff_dict : dict
         A dictionary containing tensors of shape (B,C,H,W)
 
-    Note
-    ----
+    Notes
+    -----
     The inverse operation to `local_gain_control_dict`.
     This function is  an analogue to polar_to_rectangular_dict for real
     valued signals. For more details, see :meth:`local_gain_release`
