### 2023-04-14 947abf6
adds example synth image to index.rst
