### 2023-04-14 bfe65c4
adds tips and tricks page
