### 2023-04-18 #97 @NickleDave
Welcome @billbrod!
Thank you for raising a presubmission issue.

The short answer is that, yes, plenoptic is in scope.

I can understand why that might not have been clear: you were looking for a category in our guide that the package fits into?

It definitely falls under the category of "high-level package" in the sense described at the top of the page here: https://www.pyopensci.org/software-peer-review/about/package-scope.html#what-types-of-packages-does-pyopensci-review

As you noted

> this package is intended for use by the vision science, machine learning, and neuroscience communities

and so reviewing the package would help us help you support scientific workflows in those domains. I am [a little familiar](https://github.com/NickleDave#visual-search-and-visual-attention) with this area, so I feel safe saying that I think I part of what's going on here is that methods are still evolving. But making it easier for other people to apply and explore these methods will facilitate that evolution. Open science ✨😼 And of course it's line with everything you (all) are doing at Flatiron.

Along the same lines: some feedback from our editorial team was that the purpose of the project was not clear. You made it clear to me here at least, but I agree that "surfacing" the core functionality on the index page and README would really help. It looks like the PR docs build you linked to will go a long way towards helping with that. I'd suggest adding similar images to the README if possible. But this is just tweaking the delivery; all the right info is there.

It looks like you are a good way towards meeting all the [requirements](https://www.pyopensci.org/software-peer-review/how-to/author-guide.html#does-your-package-meet-packaging-requirements).
One thing I noticed is that you are using a setup.py. It's not quite clear to me at first glance if you would be able to instead use a modern [pyproject.toml with declarative metadata](https://www.pyopensci.org/python-package-guide/package-structure-code/pyproject-toml-python-package-metadata.html). I'd strongly suggest doing so if you can. I'm also a little surprised to see some of the modules listed as packages--not sure if that's necessary? Just some quick feedback I noticed on an initial look-through.

In short: please proceed with a full submission if you would like, and we will be happy to provide you with a review. We can close this once you open that submission. I'm also happy to answer more questions here.
