### 2023-04-18 #97 @NickleDave
One more comment:
as far as related packages you might consider including links to the texforms work from @brialorelle @ArturoDeza e.g. https://github.com/brialorelle/texformgen and https://github.com/ArturoDeza/Fast-Texforms (that IIUC are an extension of metamers)
