### 2023-04-18 #97 @billbrod
Yes, I thought that since it's a "high level package", it had to fall within one of the supported domain areas. That's not the case right now? Then there is no domain restriction?

But I'm glad to hear plenoptic is in scope! Yes, the purpose definitely needs to be more clear. I've got some open PRs to finish and merge first (including the docs one I linked and another about updating the README), but then I'll proceed with a full submission, probably around the time of our presentation at [VSS](https://www.visionsciences.org/presentation/?id=5822).

To answer your comments:
1. I'm unclear about the distinctions between the different ways of packaging python packages, and setup.py was just the one I was already familiar with. I did just come across your packaging guideline recently, and am planning on reading it over. If pyproject.toml is strongly recommended, I can switch to that.
2. I don't think the modules listed under packages in setup.py are necessary: again, legacy of what I did before (they may have been necessary in [pyrtools](https://github.com/LabForComputationalVision/pyrtools/), the first library I packaged, which has some C code). Do you have a good explainer for what listed them under packages does? Again, it's unclear to me.
3. For related packages, I'm unclear how much they have to be "packages" rather than "research artifacts", for lack of a better term. There are plenty of related github repos of code that are associated with research papers (including those), but in my mind they're not packages, because they're meant to reproduce a specific paper, rather than work generally, are maintained and pip-installable, etc. Is that not a useful distinction? I agree we should definitely link them as related work somewhere in the docs, though.
