### 2023-04-18 #97 @NickleDave
> Then there is no domain restriction?

"High level package" is sort of the top-level criteria; there's no restriction on research domain (e.g., neuroscience). The categories guide what we focus on in terms of open science but there is no requirement that a package fit neatly into only one of those categories.

> I'll proceed with a full submission, probably around the time of our presentation at [VSS](https://www.visionsciences.org/presentation/?id=5822).

Perfect. By the way, in case you didn't see / weren't aware, we partner with JOSS

> Do you have a good explainer for what listed them under packages does?

I think the point of being able to list multiple things in `packages` is when you have a "monorepo" type situation as shown in this SO post:
https://stackoverflow.com/questions/42859852/distributing-multiple-packages-using-setup-py-for-python
In your case it works because a single module can technically be a package, I think. But that is a guess.
I would try installing + building with pyproject.toml, and if that works, avoid worrying about `setuptools.setup` unless you need to for a package like pyrtools.

> Is that not a useful distinction?

Yes, sorry I should have checked whether the Python project was a full-fledged package. If not then I agree that citing in a "Related works" page of the docs is appropriate here.

Look forward to your submission, thank you!
