### 2023-04-19 #97 @NickleDave
> Glad I opened this presubmission inquiry!

Us too 🙂 How we describe scope is still evolving [as we get back into gear](https://www.pyopensci.org/blog/new-executive-director-leah-wasser)--if you have feedback on how we could make it clearer that in general we're happy to review higher level packages , happy to hear it

> Should we close this issue now or after the submission?
We can close now but please reference this issue (by issue # I mean) when you open the submission
