### 2023-04-19 #97 @billbrod
I was confused about scope since [they say](https://www.pyopensci.org/software-peer-review/about/package-scope.html#is-your-package-in-scope-for-pyopensci-review): "pyOpenSci reviews packages that fall within a list of specified categories and domains. Packages must also meet our technical scope requirements." So I then looked through the `Package categories` and `Domain areas` lists, and assumed that, if my package didn't fall under one of those categories, it was not in scope. To change it maybe change that line I quoted to say "pyOpenSci reviews packages that fall within a list of specified categories and domains, as well as high-level packages in other areas. Packages must also meet our technical scope requirements."  or something similar? Or, at the top of `Domain areas`, add a note along the lines of "high-level scientific packages in other domains are also encouraged to submit, but these domain areas are particularly well-supported / we have existing ties / something" (I'm not clear on what's different about the domains listed on the website, to be honest).

>  We can close now but please reference this issue (by issue # I mean) when you open the submission

Will do!
